+++ src/export/FileWriter.js
@@ -1,5 +1,4 @@
 "use babel";
-"use strict";
 
 import { remote } from 'electron';
 import path from 'path';